[CodeQuality] Rename and refocus rule to flip void-method with(callback) to willReturnCallback
