--- rules-tests/CodeQuality/Rector/Class_/RemoveReturnFromVoidMethodMockCallbackRector/Fixture/return_throw_and_null.php.inc
@@ -1,53 +1,0 @@
-<?php
-
-namespace Rector\PHPUnit\Tests\CodeQuality\Rector\Class_\RemoveReturnFromVoidMethodMockCallbackRector\Fixture;
-
-use PHPUnit\Framework\TestCase;
-use Rector\PHPUnit\Tests\CodeQuality\Rector\Class_\RemoveReturnFromVoidMethodMockCallbackRector\Source\SomeEntityManager;
-
-final class ReturnThrowAndNull extends TestCase
-{
-    public function test($matcher): void
-    {
-        $this->createMock(SomeEntityManager::class)
-            ->method('persist')
-            ->willReturnCallback(function ($entity) use ($matcher) {
-                if (1 === $matcher->numberOfInvocations()) {
-                    return null;
-                }
-
-                if (2 === $matcher->numberOfInvocations()) {
-                    return throw new \RuntimeException('boom');
-                }
-            });
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\PHPUnit\Tests\CodeQuality\Rector\Class_\RemoveReturnFromVoidMethodMockCallbackRector\Fixture;
-
-use PHPUnit\Framework\TestCase;
-use Rector\PHPUnit\Tests\CodeQuality\Rector\Class_\RemoveReturnFromVoidMethodMockCallbackRector\Source\SomeEntityManager;
-
-final class ReturnThrowAndNull extends TestCase
-{
-    public function test($matcher): void
-    {
-        $this->createMock(SomeEntityManager::class)
-            ->method('persist')
-            ->willReturnCallback(function ($entity) use ($matcher): void {
-                if (1 === $matcher->numberOfInvocations()) {
-                    return;
-                }
-
-                if (2 === $matcher->numberOfInvocations()) {
-                    throw new \RuntimeException('boom');
-                }
-            });
-    }
-}
-
-?>
