--- rules-tests/CodeQuality/Rector/Class_/RemoveReturnFromVoidMethodMockCallbackRector/Fixture/will_return_callback_void.php.inc
@@ -1,43 +1,0 @@
-<?php
-
-namespace Rector\PHPUnit\Tests\CodeQuality\Rector\Class_\RemoveReturnFromVoidMethodMockCallbackRector\Fixture;
-
-use PHPUnit\Framework\TestCase;
-use Rector\PHPUnit\Tests\CodeQuality\Rector\Class_\RemoveReturnFromVoidMethodMockCallbackRector\Source\SomeEntityManager;
-
-final class WillReturnCallbackVoid extends TestCase
-{
-    public function test($value): void
-    {
-        $this->createMock(SomeEntityManager::class)
-            ->method('persist')
-            ->willReturnCallback(function ($entity) {
-                echo $entity;
-
-                return true;
-            });
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\PHPUnit\Tests\CodeQuality\Rector\Class_\RemoveReturnFromVoidMethodMockCallbackRector\Fixture;
-
-use PHPUnit\Framework\TestCase;
-use Rector\PHPUnit\Tests\CodeQuality\Rector\Class_\RemoveReturnFromVoidMethodMockCallbackRector\Source\SomeEntityManager;
-
-final class WillReturnCallbackVoid extends TestCase
-{
-    public function test($value): void
-    {
-        $this->createMock(SomeEntityManager::class)
-            ->method('persist')
-            ->willReturnCallback(function ($entity): void {
-                echo $entity;
-            });
-    }
-}
-
-?>
