--- rules-tests/CodeQuality/Rector/Class_/RemoveReturnFromVoidMethodMockCallbackRector/config/configured_rule.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\PHPUnit\CodeQuality\Rector\Class_\RemoveReturnFromVoidMethodMockCallbackRector;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(RemoveReturnFromVoidMethodMockCallbackRector::class);
-};
