+++ rules-tests/CodeQuality/Rector/Class_/VoidMethodWithCallbackToWillReturnCallbackRector/config/configured_rule.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\Config\RectorConfig;
+use Rector\PHPUnit\CodeQuality\Rector\Class_\VoidMethodWithCallbackToWillReturnCallbackRector;
+
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->rule(VoidMethodWithCallbackToWillReturnCallbackRector::class);
+};
